move source resolving and conflict validation into DatasetDefinition
